[MISC] Move is_constexpr test to basic type traits test.
